requirements.txt
Django==3.1
djangorestframework==3.11
psycopg2==2.8.5
django-mptt==0.11.0
pandas==1.0.3
natsort==7.0.1

flake8==3.8.3
